Extract text tile buttons, to reduce cycles
