+++ src/plugins/shared-variables/slate/variables-plugin.tsx
@@ -6,29 +6,18 @@ import {
   RenderElementProps,
   useSerializing,
   useSelected,
-  Transforms,
   CustomEditor,
   BaseElement,
   CustomElement,
   registerElementComponent,
 } from "@concord-consortium/slate-editor";
+import { action, autorun, computed, makeObservable, observable } from "mobx";
 import { getType } from "mobx-state-tree";
+import { observer } from "mobx-react";
 import { VariableChip, VariableType } from "@concord-consortium/diagram-view";
-import { variableBuckets } from "../shared-variables-utils";
 import { TextContentModelType } from "../../../models/tiles/text/text-content";
-import { TextContentModelContext } from "../../../components/tiles/text/text-content-context";
-import { IButtonDefProps } from "../../../models/tiles/text/text-plugin-info";
-import { useNewVariableDialog } from "../dialog/use-new-variable-dialog";
-import { TextToolbarButton } from "../../../components/tiles/text/text-toolbar-button";
 import { TextPluginsContext } from "../../../components/tiles/text/text-plugins-context";
 
-import AddVariableChipIcon from "../assets/add-variable-chip-icon.svg";
-import InsertVariableChipIcon from "../assets/insert-variable-chip-icon.svg";
-import VariableEditorIcon from "../assets/variable-editor-icon.svg";
-import { useInsertVariableDialog } from "../dialog/use-insert-variable-dialog";
-import { useEditVariableDialog } from "../dialog/use-edit-variable-dialog";
-import { observer } from "mobx-react";
-import { action, autorun, computed, makeObservable, observable } from "mobx";
 import { DEBUG_SHARED_MODELS } from "../../../lib/debug";
 import { SharedVariables, SharedVariablesType } from "../shared-variables";
 
@@ -142,43 +131,6 @@ export const isVariableElement = (element: CustomElement): element is VariableEl
   return element.type === kVariableFormat;
 };
 
-export const findSelectedVariable = (selectedElements: any, variables: VariableType[]) => {
-  let selected = undefined;
-    // FIXME: The editor.selectedElements claims that it returns a BaseElement[] but it really returns a NodeEntry
-    // which is a list of pairs. [Node, Path]
-    // https://docs.slatejs.org/api/nodes/node-entry
-    // There's some weirdness below to work around that, but
-    // we should either update the return type our slate lib or just return the BaseElement list.
-  selectedElements?.forEach((selectedItem: any) => {
-    const baseElement = (selectedItem as any)[0];
-    if (isVariableElement(baseElement)) {
-      const {reference} = baseElement;
-      selected = variables.find(v => v.id === reference);
-    }
-  });
-  return selected;
-};
-
-export const insertTextVariable = (variable: VariableType, editor?: Editor) => {
-  if (!editor) {
-    console.warn("inserting variable but there is no editor");
-    return;
-  }
-  const reference = variable.id;
-  const varElt: VariableElement = { type: kVariableFormat, reference, children: [{text: "" }]};
-  Transforms.insertNodes(editor, varElt);
-};
-
-export const insertTextVariables = (variables: VariableType[], editor?: Editor) => {
-  if (!editor) {
-    console.warn("inserting variable but there is no editor");
-    return;
-  }
-  variables.forEach((variable) =>{
-    insertTextVariable(variable, editor);
- });
-};
-
 export const VariableComponent = observer(function({ attributes, children, element }: RenderElementProps) {
   const plugins = useContext(TextPluginsContext);
   // FIXME: need a const for the plugin name
@@ -228,116 +180,3 @@ export function withVariables(editor: Editor, textContent: TextContentModelType)
 
   return editor;
 }
-
-//   {
-//     iconName: "new-variable",
-//     Icon: AddVariableChipIcon,
-//     toolTip: "New Variable",
-//     buttonEnabled: () => true,
-//     command(editor) {
-//       // do something like useNewVariableDialog
-//     },
-//   },
-export const NewVariableTextButton = observer(function NewVariableTextButton(
-    {editor, pluginInstance}: IButtonDefProps) {
-
-  // TODO: perhaps the pluginInstance is undefined?
-  const variablesPlugin = pluginInstance as VariablesPlugin;
-
-  const isSelected = false;
-
-  const sharedModel = variablesPlugin.sharedModel;
-
-  const enabled = !!sharedModel;
-
-  const highlightedText = (editor && editor.selection) ? Editor.string(editor, editor.selection) : "";
-  const namePrefill = highlightedText;
-  const [showDialog] = useNewVariableDialog({
-    addVariable(variable) {
-      insertTextVariable(variable, editor);
-    },
-    sharedModel, namePrefill});
-  const handleClick = (event: React.MouseEvent) => {
-    event.preventDefault();
-    showDialog();
-  };
-  return (
-    <TextToolbarButton iconName="new-variable" Icon={AddVariableChipIcon}
-      tooltip={"New Variable"}  enabled={enabled} isSelected={isSelected}
-      onClick={handleClick} />
-  );
-});
-
-//   {
-//     iconName: "insert-variable",
-//     Icon: InsertVariableChipIcon,
-//     toolTip: "Insert Variable",
-//     buttonEnabled: () => true,
-//     command(editor) {
-//       // do something like useInsertVariableDialog
-//     },
-//   },
-export const InsertVariableTextButton = observer(function InsertVariableTextButton(
-     {editor, pluginInstance}: IButtonDefProps) {
-  // TODO: perhaps the pluginInstance is undefined?
-  const variablesPlugin = pluginInstance as VariablesPlugin;
-
-
-  const isSelected = false;
-  const textContent = useContext(TextContentModelContext);
-  const sharedModel = variablesPlugin.sharedModel;
-  const enabled = !!sharedModel;
-
-  const { selfVariables, otherVariables, unusedVariables } = variableBuckets(textContent, sharedModel);
-
-  const [showDialog] = useInsertVariableDialog({
-    Icon: InsertVariableChipIcon,
-    insertVariables(variables){
-      insertTextVariables(variables, editor);
-    },
-    otherVariables, selfVariables, unusedVariables });
-  const handleClick = (event: React.MouseEvent) => {
-    event.preventDefault();
-    showDialog();
-  };
-  return (
-    <TextToolbarButton iconName="insert-variable" Icon={InsertVariableChipIcon}
-      tooltip={"Insert Variable"}  enabled={enabled} isSelected={isSelected}
-      onClick={handleClick} />
-  );
-});
-
-//   {
-//     iconName: "edit-variable",
-//     Icon: VariableEditorIcon,
-//     toolTip: "Edit Variable",
-//     buttonEnabled: shouldShowEditVariableButton,
-//     command(editor) {
-//       // do something like useEditVariableDialog
-//     },
-//   }
-export const EditVariableTextButton = observer(function EditVariableTextButton(
-    {editor, pluginInstance}: IButtonDefProps) {
-  // TODO: perhaps the pluginInstance is undefined?
-  const variablesPlugin = pluginInstance as VariablesPlugin;
-
-  const isSelected = false;
-
-  const selectedElements = editor?.selectedElements();
-  const variables = variablesPlugin.variables;
-  const hasVariable = editor?.isElementActive(kVariableFormat);
-  const selectedVariable = hasVariable ? findSelectedVariable(selectedElements, variables) : undefined;
-  const enabled = !!selectedVariable;
-
-  const [showDialog] = useEditVariableDialog({variable: selectedVariable});
-  const handleClick = (event: React.MouseEvent) => {
-    event.preventDefault();
-    showDialog();
-  };
-
-  return (
-    <TextToolbarButton iconName="edit-variable" Icon={VariableEditorIcon}
-      tooltip={"Edit Variable"} enabled={enabled} isSelected={isSelected}
-      onClick={handleClick} />
-  );
-});
